[ENH] Makes datasets module and adds tests
