--- netneurotools/tests/__init__.py
@@ -1,5 +1,0 @@
-__all__ = [
-    'make_correlated_xy'
-]
-
-from .utils import make_correlated_xy
